+++ api/schemas.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class FinancialStatementRequestVM(BaseModel):
+    account_id: str
+    loan_reference: str
