+++ requirements.txt
@@ -9,3 +9,4 @@ gunicorn==19.9.0
 uvloop
 httptools
 requests
+python-dotenv
